//saveAlbum Created $albumID and included if/else  statements to associate with other parameters (title, year, artist, genre)
//albums
.css and .js documents are stored locally and their href were changed accordingly

//albumDetails
- .css and .js documents are stored locally and their href were changed accordingly;
- added a hidden input for albumID (it's inputed automatically and associated with the rest of the info)

+++ albums.php
@@ -3,7 +3,7 @@
     <head>
         <meta charset="UTF-8">
         <title>Albums</title>
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css" integrity="sha384-rwoIResjU2yc3z8GV/NPeZWAv56rSmLldC3R/AZzGRnGxQQKnKkoFVhFQhNUwEyJ" crossorigin="anonymous">
+        <link rel="stylesheet" href="css/bootstrap.min.css">
     </head>
     <body>
         <h1>Albums</h1>
@@ -21,25 +21,26 @@
 
         $conn = null;
 
-        echo '<table class="table table-striped table-hover">
-              <tr><th>Album Title</th>
+        echo '<table class="table table-striped table-hover"><tr>
+                  <th>Album Title</th>
                   <th>Year</th>
                   <th>Artist</th>
                   <th>Genre</th>
                   <th>Edit</th>
                   <th>Delete</th></tr>';
 
-        foreach($albums as $album)
-        {
+        foreach($albums as $album) {
             echo '<tr><td>'.$album['title'].'</td>
                       <td>'.$album['year'].'</td>
                       <td>'.$album['artist'].'</td>
                       <td>'.$album['genre'].'</td>
                       <th><a href="albumDetails.php?albumID='.$album['albumID'].'" class="btn btn-primary">Edit</a></th>
-                      <td><a href="deleteAlbum.php?albumID='.$album['albumID'].'" class="btn btn-danger">Delete</a></td></tr>';
+                      <td><a href="deleteAlbum.php?albumID='.$album['albumID'].'" class="btn btn-danger confirmation">Delete</a></td></tr>';
         }
         echo '</table>';
     ?>
     </body>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+    <script src="js/jquery-3.2.1.min.js"></script>
+    <script src="js/bootstrap.min.js"></script>
+    <script src="js/app.js"></script> <!--Custom js--!>
 </html>
